# Optimized parseQuery helper function