Optimized parseQuery function. Less code and faster. Please see the rough performance test result below.

|iterate|old parseQuery(ms)|new parseQuery(ms)|
|---|---|---|
|10|1|1|
|10^2|6|2|
|10^3|10|7|
|10^4|56|48|
|10^5|534|268|
|10^6|4820|2985|

```
const selectors = ['div.hello.world', 'p#id', 'ul#hello.world', 'footer.footer.top', 'button.button#submitButton', '.usa.world'];

function test1(iterate) {
	const initTime = Date.now();
	const record = [];
	for (let i = 0; i < iterate; i++) {
		selectors.forEach(selector => {
			record.push(parseQuery(selector));
		});
	}
	const closeTime = Date.now();
	console.log(closeTime - initTime, record.length);
}

function test2(iterate) {
	const initTime = Date.now();
	const record = [];
	for (let i = 0; i < iterate; i++) {
		selectors.forEach(selector => {
			record.push(parseDOMSelector(selector));
		});
	}
	const closeTime = Date.now();
	console.log(closeTime - initTime, record.length);
}

const iterate = 1000000;
test1(iterate);
test2(iterate);
```